# perf: inline hot path and remove `IsInstanceOf` usage