- Add `[MethodImpl(MethodImplOptions.AggressiveInlining)]` to `CalculateWidth`, I assumed the JIT would do this but perhaps not
- Use `SyntaxKind` in `IsNodeIgnored`

### Benchmarks
Benchmarks appear to show a suprising speedup, although I don't know if this is accurate.
#### Before

| Method                | Mean    | Error    | StdDev   | Median  | Gen0       | Gen1       | Gen2       | Allocated |
|---------------------- |--------:|---------:|---------:|--------:|-----------:|-----------:|-----------:|----------:|
| Default_CodeFormatter | 1.637 s | 0.0329 s | 0.0969 s | 1.598 s | 52000.0000 | 31000.0000 | 10000.0000 | 420.54 MB |
| Method                | Mean    | Error    | StdDev   | Gen0       | Gen1       | Gen2      | Allocated |
|---------------------- |--------:|---------:|---------:|-----------:|-----------:|----------:|----------:|
| Default_CodeFormatter | 1.571 s | 0.0308 s | 0.0822 s | 50000.0000 | 29000.0000 | 9000.0000 | 419.39 MB |
`


#### After

| Method                | Mean    | Error    | StdDev   | Gen0       | Gen1       | Gen2       | Allocated |
|---------------------- |--------:|---------:|---------:|-----------:|-----------:|-----------:|----------:|
| Default_CodeFormatter | 1.399 s | 0.0183 s | 0.0171 s | 52000.0000 | 31000.0000 | 10000.0000 | 420.36 MB |
| Method                | Mean    | Error    | StdDev   | Gen0       | Gen1       | Gen2       | Allocated |
|---------------------- |--------:|---------:|---------:|-----------:|-----------:|-----------:|----------:|
| Default_CodeFormatter | 1.406 s | 0.0066 s | 0.0059 s | 52000.0000 | 31000.0000 | 10000.0000 | 420.52 MB |
| Method                | Mean    | Error    | StdDev   | Median  | Gen0       | Gen1       | Gen2       | Allocated |
|---------------------- |--------:|---------:|---------:|--------:|-----------:|-----------:|-----------:|----------:|
| Default_CodeFormatter | 1.457 s | 0.0285 s | 0.0740 s | 1.425 s | 52000.0000 | 31000.0000 | 10000.0000 | 420.38 MB |